python-orchestrator/requirements.txt
google-genai
openpyxl
playwright
pydantic
python-dotenv
